pyproject.toml
[project]
name = "serverapi"
version = "0.1.0"
requires-python = ">=3.10,<4"
dependencies = [
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "tortoise-orm (>=0.24.0,<0.25.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "ujson (>=5.10.0,<6.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "mcstatus (>=11.1.1,<12.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aioboto3 (>=13.4.0,<14.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "meilisearch (>=0.34.0,<0.35.0)",
    "bcrypt==4.0.1",
    "tomlkit>=0.13.2",
    "tzdata>=2025.2",
    "argparse>=1.4.0",
]


[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]


[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[dependency-groups]
dev = ["aerich>=0.8.2"]
